alterações da sala ye
