+++ src/App.js
@@ -31,7 +31,7 @@ import routes from "routes";
 import { useSoftUIController, setMiniSidenav, setOpenConfigurator } from "context";
 
 // Images
-import brand from "assets/images/logo-ct.png";
+import LogoIcon from "imgs/Logo.png";
 
 export default function App() {
   const [controller, dispatch] = useSoftUIController();
@@ -112,7 +112,7 @@ export default function App() {
       onClick={handleConfiguratorOpen}
     >
       <Icon fontSize="default" color="inherit">
-        settings
+        Settings
       </Icon>
     </SoftBox>
   );
@@ -125,7 +125,7 @@ export default function App() {
           <>
             <Sidenav
               color={sidenavColor}
-              brand={brand}
+              brand={LogoIcon}
               brandName="Soft UI Dashboard"
               routes={routes}
               onMouseEnter={handleOnMouseEnter}
@@ -149,8 +149,8 @@ export default function App() {
         <>
           <Sidenav
             color={sidenavColor}
-            brand={brand}
-            brandName="Soft UI Dashboard"
+            brand={LogoIcon}
+            brandName="ComParty"
             routes={routes}
             onMouseEnter={handleOnMouseEnter}
             onMouseLeave={handleOnMouseLeave}
