--- src/layouts/billing/components/Bill/index.js
@@ -1,94 +1,0 @@
-// prop-types is a library for typechecking of props
-import PropTypes from "prop-types";
-
-// @mui material components
-import Icon from "@mui/material/Icon";
-
-// Soft UI Dashboard React components
-import SoftBox from "components/SoftBox";
-import SoftTypography from "components/SoftTypography";
-import SoftButton from "components/SoftButton";
-
-function Bill({ name, company, email, vat, noGutter }) {
-  return (
-    <SoftBox
-      component="li"
-      display="flex"
-      justifyContent="space-between"
-      alignItems="flex-start"
-      bgColor="grey-100"
-      borderRadius="lg"
-      p={3}
-      mb={noGutter ? 0 : 1}
-      mt={2}
-    >
-      <SoftBox width="100%" display="flex" flexDirection="column">
-        <SoftBox
-          display="flex"
-          justifyContent="space-between"
-          alignItems={{ xs: "flex-start", sm: "center" }}
-          flexDirection={{ xs: "column", sm: "row" }}
-          mb={2}
-        >
-          <SoftTypography variant="button" fontWeight="medium" textTransform="capitalize">
-            {name}
-          </SoftTypography>
-
-          <SoftBox
-            display="flex"
-            alignItems="center"
-            mt={{ xs: 2, sm: 0 }}
-            ml={{ xs: -1.5, sm: 0 }}
-          >
-            <SoftBox mr={1}>
-              <SoftButton variant="text" color="error">
-                <Icon>delete</Icon>&nbsp;delete
-              </SoftButton>
-            </SoftBox>
-            <SoftButton variant="text" color="dark">
-              <Icon>edit</Icon>&nbsp;edit
-            </SoftButton>
-          </SoftBox>
-        </SoftBox>
-        <SoftBox mb={1} lineHeight={0}>
-          <SoftTypography variant="caption" color="text">
-            Company Name:&nbsp;&nbsp;&nbsp;
-            <SoftTypography variant="caption" fontWeight="medium" textTransform="capitalize">
-              {company}
-            </SoftTypography>
-          </SoftTypography>
-        </SoftBox>
-        <SoftBox mb={1} lineHeight={0}>
-          <SoftTypography variant="caption" color="text">
-            Email Address:&nbsp;&nbsp;&nbsp;
-            <SoftTypography variant="caption" fontWeight="medium">
-              {email}
-            </SoftTypography>
-          </SoftTypography>
-        </SoftBox>
-        <SoftTypography variant="caption" color="text">
-          VAT Number:&nbsp;&nbsp;&nbsp;
-          <SoftTypography variant="caption" fontWeight="medium">
-            {vat}
-          </SoftTypography>
-        </SoftTypography>
-      </SoftBox>
-    </SoftBox>
-  );
-}
-
-// Setting default values for the props of Bill
-Bill.defaultProps = {
-  noGutter: false,
-};
-
-// Typechecking props for the Bill
-Bill.propTypes = {
-  name: PropTypes.string.isRequired,
-  company: PropTypes.string.isRequired,
-  email: PropTypes.string.isRequired,
-  vat: PropTypes.string.isRequired,
-  noGutter: PropTypes.bool,
-};
-
-export default Bill;
