--- src/layouts/billing/components/BillingInformation/index.js
@@ -1,46 +1,0 @@
-// @mui material components
-import Card from "@mui/material/Card";
-
-// Soft UI Dashboard React components
-import SoftBox from "components/SoftBox";
-import SoftTypography from "components/SoftTypography";
-
-// Billing page components
-import Bill from "layouts/billing/components/Bill";
-
-function BillingInformation() {
-  return (
-    <Card id="delete-account">
-      <SoftBox pt={3} px={2}>
-        <SoftTypography variant="h6" fontWeight="medium">
-          Billing Information
-        </SoftTypography>
-      </SoftBox>
-      <SoftBox pt={1} pb={2} px={2}>
-        <SoftBox component="ul" display="flex" flexDirection="column" p={0} m={0}>
-          <Bill
-            name="oliver liam"
-            company="viking burrito"
-            email="oliver@burrito.com"
-            vat="FRB1235476"
-          />
-          <Bill
-            name="lucas harper"
-            company="stone tech zone"
-            email="lucas@stone-tech.com"
-            vat="FRB1235476"
-          />
-          <Bill
-            name="ethan james"
-            company="fiber notion"
-            email="ethan@fiber.com"
-            vat="FRB1235476"
-            noGutter
-          />
-        </SoftBox>
-      </SoftBox>
-    </Card>
-  );
-}
-
-export default BillingInformation;
