+++ src/layouts/room/index.js
@@ -9,19 +9,15 @@ import DashboardLayout from "examples/LayoutContainers/DashboardLayout";
 import DashboardNavbar from "examples/Navbars/DashboardNavbar";
 
 // Billing page components
-import BillingInformation from "layouts/billing/components/BillingInformation";
-import Transactions from "layouts/billing/components/Transactions";
+import Transactions from "layouts/room/components/Transactions";
 
-function Billing() {
+function Room() {
   return (
     <DashboardLayout>
       <DashboardNavbar />
       <SoftBox mt={4}>
         <SoftBox my={3}>
           <Grid container spacing={3}>
-            <Grid item xs={12} md={7}>
-              <BillingInformation />
-            </Grid>
             <Grid item xs={12} md={5}>
               <Transactions />
             </Grid>
@@ -32,4 +28,4 @@ function Billing() {
   );
 }
 
-export default Billing;
+export default Room;
